+++ kitchen.py
@@ -1,0 +1,5 @@
+from location import *
+
+class Kitchen(Location):
+  def __init__(self, name, up, right, down, left):
+    super().__init__(name, up, right, down, left)
