--- locations/world.py
@@ -1,7 +1,0 @@
-from house import *
-def move():
-  pass
-
-world = {
-  "House": House("House","Kitchen")
-}
