+++ pyproject.toml
@@ -9,6 +9,7 @@ python = ">=3.10.0"
 pygame = "^2.1.2"
 dearpygui = "^1.6.2"
 scripts = "^3.0"
+django-locations-base = "^0.3.4"
 
 [tool.pyright]
 # https://github.com/microsoft/pyright/blob/main/docs/configuration.md
